requirements.txt
streamlit
pandas
seaborn
jsons
trueskill
pyyaml
plotly
pydot
python-dateutil
pytest